+++ xenon-common/src/main/java/com/vmware/xenon/common/http/netty/NettyChannelPool.java
@@ -637,6 +637,13 @@ public class NettyChannelPool {
                 isClose = isClose || !ch.isOpen() || !context.isValid();
             } else {
                 isClose = isClose || !ch.isWritable() || !ch.isOpen();
+
+                // When a client-side timeout occurs, the channel must be closed rather than reused
+                // since the server may still send a response on this channel even though the
+                // operation has been failed from the caller's perspective. If the channel still
+                // has a pending operation, then close it.
+                Operation pendingOp = ch.attr(NettyChannelContext.OPERATION_KEY).getAndSet(null);
+                isClose = isClose || (pendingOp != null);
             }
         }